Convert the cookiecutter to poetry and fix the formatting.
