+++ python/pytest-selfie/tests/test_pytest_selfie.py
@@ -8,15 +8,14 @@ def test_bar_fixture(pytester):
     """)
 
     # run pytest with the following cmd args
-    result = pytester.runpytest(
-        '--foo=europython2015',
-        '-v'
-    )
+    result = pytester.runpytest("--foo=europython2015", "-v")
 
     # fnmatch_lines does an assertion internally
-    result.stdout.fnmatch_lines([
-        '*::test_sth PASSED*',
-    ])
+    result.stdout.fnmatch_lines(
+        [
+            "*::test_sth PASSED*",
+        ]
+    )
 
     # make sure that we get a '0' exit code for the testsuite
     assert result.ret == 0
@@ -24,13 +23,15 @@ def test_bar_fixture(pytester):
 
 def test_help_message(pytester):
     result = pytester.runpytest(
-        '--help',
+        "--help",
     )
     # fnmatch_lines does an assertion internally
-    result.stdout.fnmatch_lines([
-        'selfie:',
-        '*--foo=DEST_FOO*Set the value for the fixture "bar".',
-    ])
+    result.stdout.fnmatch_lines(
+        [
+            "selfie:",
+            '*--foo=DEST_FOO*Set the value for the fixture "bar".',
+        ]
+    )
 
 
 def test_hello_ini_setting(pytester):
@@ -50,12 +51,14 @@ def test_hello_ini_setting(pytester):
             assert hello == 'world'
     """)
 
-    result = pytester.runpytest('-v')
+    result = pytester.runpytest("-v")
 
     # fnmatch_lines does an assertion internally
-    result.stdout.fnmatch_lines([
-        '*::test_hello_world PASSED*',
-    ])
+    result.stdout.fnmatch_lines(
+        [
+            "*::test_hello_world PASSED*",
+        ]
+    )
 
     # make sure that we get a '0' exit code for the testsuite
     assert result.ret == 0
